+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Passing simulationBehavior to compatibilityLayer so it calls simulation for the default behavior.
+
 ## [1.13.3] - 2020-08-06
 
 ### Fixed